@@ -1,8 +1,8 @@
 package com.mrhabibi.autonomousdialog.utils;
 
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
 
 import com.mrhabibi.autonomousdialog.wrapper.DialogWrapper;
 
