@@ -2,7 +2,8 @@ package com.mrhabibi.autonomousdialog.wrapper;
 
 import android.content.DialogInterface;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
+
+import androidx.annotation.Nullable;
 
 import com.mrhabibi.autonomousdialog.DialogBuilder;
 
