@@ -2,9 +2,10 @@ package com.mrhabibi.autonomousdialog.wrapper;
 
 import android.content.DialogInterface;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v4.app.Fragment;
-import android.support.v7.app.AlertDialog;
+
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AlertDialog;
+import androidx.fragment.app.Fragment;
 
 import com.mrhabibi.autonomousdialog.DialogBuilder;
 import com.mrhabibi.autonomousdialog.DialogCallback;
