@@ -1,7 +1,8 @@
 package com.mrhabibi.autonomousdialog.sample;
 
 import android.os.Bundle;
-import android.support.annotation.Nullable;
+
+import androidx.annotation.Nullable;
 
 import com.mrhabibi.autonomousdialog.DialogActivity;
 
