@@ -1,10 +1,11 @@
 package com.mrhabibi.autonomousdialog.utils;
 
 import android.content.Context;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
 import android.util.Log;
 import android.util.TypedValue;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import com.mrhabibi.autonomousdialog.AutonomousDialog;
 
