@@ -1,10 +1,11 @@
 package com.mrhabibi.autonomousdialog.wrapper;
 
 import android.os.Bundle;
-import android.support.annotation.NonNull;
-import android.support.annotation.Nullable;
 import android.view.View;
 import android.widget.FrameLayout;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
 
 import com.mrhabibi.autonomousdialog.DialogBuilder;
 import com.mrhabibi.autonomousdialog.DialogCallback;
