@@ -1,9 +1,10 @@
 package com.mrhabibi.autonomousdialog.sample;
 
-import android.support.v7.app.AppCompatActivity;
 import android.os.Bundle;
 import android.view.View;
 import android.widget.Button;
+
+import androidx.appcompat.app.AppCompatActivity;
 
 import com.mrhabibi.autonomousdialog.AutonomousDialog;
 import com.mrhabibi.autonomousdialog.wrapper.BasicDialogWrapper;
